Corrected the shebang in `pre-commit`
`pre-commit` uses a bashism:  the `$()` syntax.  This means it is not compatible with POSIX `sh`;  and it (accordingly) should use a `bash`y shebang, for maximum portability.

Using `/bin/sh` as the shebang should guarantee failure on Debian, Ubuntu, and several other *nix OSes that do not symlink `/bin/sh` to `/usr/bin/bash`.